all route functions in routes.py
